Write thumbnail in one pass with atomic rename
Previously, the thumbnail was written to /tmp, read back, re-written to
/tmp and finally copied to the destination folder. This lead to many
problems, especially with certain FUSE filesystems.

Instead, write the thumbnail in a single pass to a temporary file in the
destination folder, and rename it when done to the destination file.
This avoids wasteful I/O and increases the chances of the rename being
atomic.

Closes: https://github.com/filcuc/thumbnailer/issues/10